examples/audit-logging/python/requirements.txt
axonflow>=0.10.1
python-dotenv>=1.0.0
openai>=1.0.0
